third_party/WebKit/LayoutTests/fast/peerconnection/RTCPeerConnection-createOffer.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
<head>
<script src="../../resources/js-test.js"></script>
</head>
<body>
<script>
description("Tests RTCPeerConnection createOffer.");

// Note: createOffer() calls in the test runner are successful if the
// voiceActivityDetection and iceRestart options are passed with a value of true
// and offerToReceiveAudio and offerToReceiveVideo are passed with a positive
// value. In all other cases, createOffer() fails in the test runner.

var pc = null;

function unexpectedCallback()
{
    testFailed('unexpectedCallback was called')
    finishJSTest();
}

function expectedCreateOfferFailed7(error)
{
    testPassed('expectedCreateOfferFailed7 called.');
    window.error = error;
    shouldBe('error.name', '"OperationError"');
    shouldBe('error.toString()', '"OperationError: TEST_ERROR"');
    finishJSTest();
}

function expectedCreateOfferFailed6(error)
{
    testPassed('expectedCreateOfferFailed6 called.');
    window.error = error;
    shouldBe('error.name', '"OperationError"');
    shouldBe('error.toString()', '"OperationError: TEST_ERROR"');
    shouldNotThrow("pc.createOffer(unexpectedCallback, expectedCreateOfferFailed7, {offerToReceiveVideo:0, offerToReceiveAudio:-1});");
}

function expectedCreateOfferFailed5(error) {
    testPassed('expectedCreateOfferFailed5 was called.');
    window.error = error;
    shouldBe('error.name', '"OperationError"');
    shouldBe('error.toString()', '"OperationError: TEST_ERROR"');
    shouldNotThrow("pc.createOffer(unexpectedCallback, expectedCreateOfferFailed6, {offerToReceiveVideo:-1, offerToReceiveAudio:0});");
}

function createOfferSucceeded2() {
    testPassed('createOfferSucceeded2 was called.');
    shouldNotThrow("pc.createOffer(unexpectedCallback, expectedCreateOfferFailed5, {offerToReceiveVideo:1, offerToReceiveAudio:0, voiceActivityDetection:false, iceRestart:true});");
}

function expectedCreateOfferFailed4(error) {
    testPassed('expectedCreateOfferFailed4 was called.');
    window.error = error;
    shouldBe('error.name', '"OperationError"');
    shouldBe('error.toString()', '"OperationError: TEST_ERROR"');
    shouldNotThrow("pc.createOffer(createOfferSucceeded2, unexpectedCallback, {iceRestart:true, offerToReceiveAudio:1, offerToReceiveVideo:1});");
}

function expectedCreateOfferFailed3(error) {
    testPassed('expectedCreateOfferFailed3 was called.');
    window.error = error;
    shouldBe('error.name', '"OperationError"');
    shouldBe('error.toString()', '"OperationError: TEST_ERROR"');
    shouldNotThrow("pc.createOffer(unexpectedCallback, expectedCreateOfferFailed4, {voiceActivityDetection:false});");
}

function expectedCreateOfferFailed2(error) {
    testPassed('expectedCreateOfferFailed2 was called.');
    window.error = error;
    shouldBe('error.name', '"OperationError"');
    shouldBe('error.toString()', '"OperationError: TEST_ERROR"');
    shouldNotThrow("pc.createOffer(unexpectedCallback, expectedCreateOfferFailed3, {});");
}

function expectedCreateOfferFailed1(error)
{
    testPassed('expectedCreateOfferFailed1 was called.');
    window.error = error;
    shouldBe('error.name', '"OperationError"');
    shouldBe('error.toString()', '"OperationError: TEST_ERROR"');
    shouldNotThrow("pc.createOffer(unexpectedCallback, expectedCreateOfferFailed2);");
}

function createOfferSucceeded1(sessionDescription)
{
    testPassed('createOfferSucceeded1 was called.');
    window.sessionDescription = sessionDescription;
    shouldBe('sessionDescription.type', '"offer"');
    shouldNotThrow('pc.createOffer(unexpectedCallback, expectedCreateOfferFailed1);');
}

function testExecutionOrderClosedConnection()
{
    var localPeerConnection = new RTCPeerConnection();
    localPeerConnection.close();
    var counter = 0;
    window.events = [];
    Promise.resolve().then(_ => window.events[counter++] = 1);
    localPeerConnection.createOffer(unexpectedCallback, error => {
        window.error = error;
        shouldBe('error.name', '"InvalidStateError"');
        shouldBe('error.toString()', '"InvalidStateError: The RTCPeerConnection\'s signalingState is \'closed\'."');
        window.events[counter++] = 2;
    });
    Promise.resolve().then(_ => {
        window.events[counter++] = 3;
        shouldBe('events', '[1,2,3]');
    });
}

shouldNotThrow('testExecutionOrderClosedConnection()');
pc = new RTCPeerConnection();
shouldNotThrow('pc.createOffer(createOfferSucceeded1, unexpectedCallback, {voiceActivityDetection:true, iceRestart:true, offerToReceiveAudio:1, offerToReceiveVideo:1});');

window.jsTestIsAsync = true;
window.successfullyParsed = true;
</script>
</body>
</html>
